+++ src/app/mypage/_components/PasswordCheck.tsx
@@ -4,6 +4,7 @@ import { useState } from 'react';
 
 import { z } from 'zod';
 
+import FormModal from '@/components/FormModal';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardTitle } from '@/components/ui/card';
 import { Form, FormControl, FormField, FormItem, FormLabel } from '@/components/ui/form';
@@ -41,6 +42,7 @@ const PasswordSchema = z
   });
 
 const PasswordCheck = () => {
+  const [open, setOpen] = useState(false);
   const [showPassword, setShowPassword] = useState({
     passwordShow: false,
     newPasswordShow: false,
@@ -68,6 +70,7 @@ const PasswordCheck = () => {
     try {
       const dto = { current_password: data.password, new_password: data.newPassword };
       await EditProfile<EditPasswordRequest>('password', dto);
+      setOpen(true);
     } catch (err) {
       if (err instanceof CustomError) {
         if (err.response && err.response.code === 400) {
@@ -83,113 +86,128 @@ const PasswordCheck = () => {
     form.reset();
   };
 
+  const handleModal = () => {
+    // router.push('/');
+  };
+
   return (
-    <Card className="border-0">
-      <CardTitle>비밀번호 변경</CardTitle>
-      <CardContent>
-        <div className="w-full">
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)}>
-              <div className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="password"
-                  render={({ field, fieldState: { error } }) => (
-                    <FormItem className="pt-4">
-                      <FormLabel>현재 비밀번호</FormLabel>
-                      <FormControl>
-                        <Input
-                          type={showPassword.passwordShow ? 'text' : 'password'}
-                          placeholder="현재 비밀번호를 입력해주세요."
-                          isError={!!error}
-                          errorText={error?.message}
-                          endAdornment={
-                            <button type="button" onClick={() => togglePassword('passwordShow')}>
-                              {showPassword.passwordShow ? (
-                                <Eye className="mr-2 h-[20px] text-neutral-200 " />
-                              ) : (
-                                <EyeOff className="mr-2 h-[20px] text-neutral-200 " />
-                              )}
-                            </button>
-                          }
-                          {...field}
-                        />
-                      </FormControl>
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="newPassword"
-                  render={({ field, fieldState: { error } }) => (
-                    <FormItem className="pt-4">
-                      <FormLabel>새 비밀번호</FormLabel>
-                      <FormControl>
-                        <Input
-                          type={showPassword.newPasswordShow ? 'text' : 'password'}
-                          placeholder="새 비밀번호를 입력해주세요."
-                          isError={!!error}
-                          errorText={error?.message}
-                          endAdornment={
-                            <button type="button" onClick={() => togglePassword('newPasswordShow')}>
-                              {showPassword.newPasswordShow ? (
-                                <Eye className="mr-2 h-[20px] text-neutral-200 " />
-                              ) : (
-                                <EyeOff className="mr-2 h-[20px] text-neutral-200 " />
-                              )}
-                            </button>
-                          }
-                          {...field}
-                        />
-                      </FormControl>
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="newPasswordCheck"
-                  render={({ field, fieldState: { error } }) => (
-                    <FormItem className="pt-4">
-                      <FormLabel>새 비밀번호 확인</FormLabel>
-                      <FormControl>
-                        <Input
-                          type={showPassword.newPasswordCheckShow ? 'text' : 'password'}
-                          placeholder="새 비밀번호를 입력해주세요."
-                          isError={!!error}
-                          errorText={error?.message}
-                          endAdornment={
-                            <button
-                              type="button"
-                              onClick={() => togglePassword('newPasswordCheckShow')}
-                            >
-                              {showPassword.newPasswordCheckShow ? (
-                                <Eye className="mr-2 h-[20px] text-neutral-200 " />
-                              ) : (
-                                <EyeOff className="mr-2 h-[20px] text-neutral-200 " />
-                              )}
-                            </button>
-                          }
-                          {...field}
-                        />
-                      </FormControl>
-                    </FormItem>
-                  )}
-                />
-              </div>
-              <Button
-                type="submit"
-                disabled={!form.formState.isValid}
-                variant="active"
-                size="lg"
-                className="mt-[56px] w-full"
-              >
-                변경
-              </Button>
-            </form>
-          </Form>
-        </div>
-      </CardContent>
-    </Card>
+    <>
+      <Card className="border-0">
+        <CardTitle>비밀번호 변경</CardTitle>
+        <CardContent>
+          <div className="w-full">
+            <Form {...form}>
+              <form onSubmit={form.handleSubmit(onSubmit)}>
+                <div className="space-y-4">
+                  <FormField
+                    control={form.control}
+                    name="password"
+                    render={({ field, fieldState: { error } }) => (
+                      <FormItem className="pt-4">
+                        <FormLabel>현재 비밀번호</FormLabel>
+                        <FormControl>
+                          <Input
+                            type={showPassword.passwordShow ? 'text' : 'password'}
+                            placeholder="현재 비밀번호를 입력해주세요."
+                            isError={!!error}
+                            errorText={error?.message}
+                            endAdornment={
+                              <button type="button" onClick={() => togglePassword('passwordShow')}>
+                                {showPassword.passwordShow ? (
+                                  <Eye className="mr-2 h-[20px] text-neutral-200 " />
+                                ) : (
+                                  <EyeOff className="mr-2 h-[20px] text-neutral-200 " />
+                                )}
+                              </button>
+                            }
+                            {...field}
+                          />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+                  <FormField
+                    control={form.control}
+                    name="newPassword"
+                    render={({ field, fieldState: { error } }) => (
+                      <FormItem className="pt-4">
+                        <FormLabel>새 비밀번호</FormLabel>
+                        <FormControl>
+                          <Input
+                            type={showPassword.newPasswordShow ? 'text' : 'password'}
+                            placeholder="새 비밀번호를 입력해주세요."
+                            isError={!!error}
+                            errorText={error?.message}
+                            endAdornment={
+                              <button
+                                type="button"
+                                onClick={() => togglePassword('newPasswordShow')}
+                              >
+                                {showPassword.newPasswordShow ? (
+                                  <Eye className="mr-2 h-[20px] text-neutral-200 " />
+                                ) : (
+                                  <EyeOff className="mr-2 h-[20px] text-neutral-200 " />
+                                )}
+                              </button>
+                            }
+                            {...field}
+                          />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+                  <FormField
+                    control={form.control}
+                    name="newPasswordCheck"
+                    render={({ field, fieldState: { error } }) => (
+                      <FormItem className="pt-4">
+                        <FormLabel>새 비밀번호 확인</FormLabel>
+                        <FormControl>
+                          <Input
+                            type={showPassword.newPasswordCheckShow ? 'text' : 'password'}
+                            placeholder="새 비밀번호를 입력해주세요."
+                            isError={!!error}
+                            errorText={error?.message}
+                            endAdornment={
+                              <button
+                                type="button"
+                                onClick={() => togglePassword('newPasswordCheckShow')}
+                              >
+                                {showPassword.newPasswordCheckShow ? (
+                                  <Eye className="mr-2 h-[20px] text-neutral-200 " />
+                                ) : (
+                                  <EyeOff className="mr-2 h-[20px] text-neutral-200 " />
+                                )}
+                              </button>
+                            }
+                            {...field}
+                          />
+                        </FormControl>
+                      </FormItem>
+                    )}
+                  />
+                </div>
+                <Button
+                  type="submit"
+                  disabled={!form.formState.isValid}
+                  variant="active"
+                  size="lg"
+                  className="mt-[56px] w-full"
+                >
+                  변경
+                </Button>
+              </form>
+            </Form>
+          </div>
+        </CardContent>
+      </Card>
+      <FormModal
+        content="비밀번호가 변경되었어요!"
+        handleModal={handleModal}
+        open={open}
+        setOpen={setOpen}
+      />
+    </>
   );
 };
